+++ packages/popup/src/components/index.js
@@ -36,6 +36,6 @@ export {default as TransactionResult} from './TransactionResult'
 export {default as CurrentNetworkDisplay} from './CurrentNetworkDisplay'
 export {default as ContactItem} from './ContactItem'
 export {default as ContactList} from './ContactList'
-export {default as GroupItem} from './GroupItem'
+export {default as AccountGroupItem} from './AccountGroupItem'
 export {default as AccountItem} from './AccountItem'
 export {default as LedgerGroupTag} from './LedgerGroupTag'
