+++ packages/popup/src/pages/AccountManagement/components/index.js
@@ -1,3 +1,3 @@
-export {default as TextNickNameComponent} from './TextNickNameComponent'
+export {default as TextNickname} from './TextNickname'
 export {default as GroupFooter} from './GroupFooter'
 export {default as AccountOperation} from './AccountOperation'
