+++ packages/popup/src/pages/AccountManagement/index.js
@@ -11,17 +11,13 @@ import {
   ConfirmPassword,
   SearchAccount,
   NoResult,
-  GroupItem,
+  AccountGroupItem,
   AccountItem,
   LedgerGroupTag,
 } from '../../components'
 import {useAccountList, useCurrentAddress} from '../../hooks/useApi'
 import {request} from '../../utils'
-import {
-  TextNickNameComponent,
-  GroupFooter,
-  AccountOperation,
-} from './components'
+import {TextNickname, GroupFooter, AccountOperation} from './components'
 const {EXPORT_SEED, EXPORT_PRIVATEKEY, SELECT_CREATE_TYPE} = ROUTES
 const {WALLET_EXPORT_ACCOUNT_GROUP, WALLET_EXPORT_ACCOUNT, ACCOUNT_GROUP_TYPE} =
   RPC_METHODS
@@ -212,12 +208,12 @@ function AccountManagement() {
               vault,
               eid: accountGroupId,
             }) => (
-              <GroupItem
+              <AccountGroupItem
                 key={accountGroupId}
                 nickname={groupNickname}
                 groupType={vault?.type}
                 GroupNameOverlay={
-                  <TextNickNameComponent
+                  <TextNickname
                     id={accountGroupId}
                     nickname={groupNickname}
                     accountGroupId={accountGroupId}
@@ -256,7 +252,7 @@ function AccountManagement() {
                       accountId={accountId}
                       accountNickname={accountNickname}
                       AccountNameOverlay={
-                        <TextNickNameComponent
+                        <TextNickname
                           nickname={accountNickname}
                           accountId={accountId}
                           updateEditedName={updateEditedName}
@@ -278,7 +274,7 @@ function AccountManagement() {
                     />
                   ),
                 )}
-              </GroupItem>
+              </AccountGroupItem>
             ),
           )
         )}
