+++ packages/popup/src/pages/Contacts/components/index.js
@@ -1,2 +1,2 @@
-export {default as ContactOperationComponent} from './ContactOperationComponent'
+export {default as ContactOperation} from './ContactOperation'
 export {default as DeleteContactModal} from './DeleteContactModal'
